+++ packages/calcite-components/src/utils/dom.ts
@@ -1,5 +1,6 @@
 // @ts-strict-ignore
 import { focusable, tabbable } from "tabbable";
+import { LitElement } from "@arcgis/lumina";
 import { IconNameOrString } from "../components/icon/interfaces";
 import { guid } from "./guid";
 import { CSS_UTILITY } from "./resources";
@@ -234,30 +235,20 @@ function visit<T = any>(node: Node, onVisit: (node: Node) => T): T {
   return visit(parentNode instanceof ShadowRoot ? parentNode.host : parentNode, onVisit);
 }
 
-/**
- * This helper returns true when an element has the descendant in question.
- *
- * @param {Element} element The starting element.
- * @param {Element} maybeDescendant The descendant.
- * @returns {boolean} The result.
- */
-export function containsCrossShadowBoundary(element: Element, maybeDescendant: Element): boolean {
-  return !!walkUpAncestry(maybeDescendant, (node) => (node === element ? true : undefined));
-}
+export type FocusableElement = SetFocusable | HTMLElement;
 
-/** An element which may contain a `setFocus` method. */
-export interface FocusableElement extends HTMLElement {
-  setFocus?: () => Promise<void>;
+export interface SetFocusable extends LitElement {
+  setFocus: (options?: FocusOptions) => Promise<void>;
 }
 
 /**
  * This helper returns true when an element has a setFocus method.
  *
  * @param {Element} el An element.
- * @returns {boolean} The result.
+ * @returns {boolean} Whether the element is focusable.
  */
-export function isCalciteFocusable(el: FocusableElement): boolean {
-  return typeof el?.setFocus === "function";
+export function isCalciteFocusable(el: FocusableElement): el is SetFocusable {
+  return typeof (el as SetFocusable)?.setFocus === "function";
 }
 
 /**
@@ -267,23 +258,27 @@ export function isCalciteFocusable(el: FocusableElement): boolean {
  * @param includeContainer When true, the container element will be considered as well. Note, this is only applicable when `setFocus` is not applicable.
  * @param strategy The focus strategy to use when finding the first focusable element. Defaults to "tabbable".
  * @param context The element invoking the focus – use when the host is focusable to short-circuit the focus call.
+ * @param options - When specified an optional object customizes the component's focusing process. When `preventScroll` is `true`, scrolling will not occur on the component.
+ *
+ * @mdn [focus(options)](https://developer.mozilla.org/en-US/docs/Web/API/HTMLElement/focus#options)
  */
 export async function focusElement(
   el: FocusableElement,
   includeContainer = false,
   strategy: "focusable" | "tabbable" = "tabbable",
   context?: HTMLElement,
+  options?: FocusOptions,
 ): Promise<void> {
   if (!el) {
     return;
   }
 
   if (isCalciteFocusable(el) && context !== el) {
-    return el.setFocus();
+    return el.setFocus(options);
   }
 
   const firstFocusFunction = strategy === "tabbable" ? focusFirstTabbable : focusFirstFocusable;
-  return firstFocusFunction(el, includeContainer);
+  return firstFocusFunction(el, includeContainer, options);
 }
 
 /**
@@ -307,9 +302,12 @@ export function getFirstTabbable(element: HTMLElement, includeContainer?: boolea
  *
  * @param {HTMLElement} element The html element containing tabbable elements.
  * @param {boolean} includeContainer When true, the container element will be considered as well.
+ * @param options - When specified an optional object customizes the component's focusing process. When `preventScroll` is `true`, scrolling will not occur on the component.
+ *
+ * @mdn [focus(options)](https://developer.mozilla.org/en-US/docs/Web/API/HTMLElement/focus#options)
  */
-export function focusFirstTabbable(element: HTMLElement, includeContainer?: boolean): void {
-  getFirstTabbable(element, includeContainer)?.focus();
+export function focusFirstTabbable(element: HTMLElement, includeContainer?: boolean, options?: FocusOptions): void {
+  getFirstTabbable(element, includeContainer)?.focus(options);
 }
 
 /**
@@ -338,8 +336,8 @@ function getFirstFocusable(element: HTMLElement, includeContainer?: boolean): HT
  *
  * @internal
  */
-function focusFirstFocusable(element: HTMLElement, includeContainer?: boolean): void {
-  getFirstFocusable(element, includeContainer)?.focus();
+function focusFirstFocusable(element: HTMLElement, includeContainer?: boolean, options?: FocusOptions): void {
+  getFirstFocusable(element, includeContainer)?.focus(options);
 }
 
 /**