+++ packages/calcite-components-react/CHANGELOG.md
@@ -3,6 +3,42 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.3.0-next.52](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.3.0-next.51...@esri/calcite-components-react@3.3.0-next.52) (2025-07-31)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.3.0-next.51](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.3.0-next.50...@esri/calcite-components-react@3.3.0-next.51) (2025-07-30)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.3.0-next.50](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.3.0-next.49...@esri/calcite-components-react@3.3.0-next.50) (2025-07-29)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.3.0-next.49](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.3.0-next.48...@esri/calcite-components-react@3.3.0-next.49) (2025-07-28)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.3.0-next.48](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.3.0-next.47...@esri/calcite-components-react@3.3.0-next.48) (2025-07-25)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.3.0-next.47](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.3.0-next.46...@esri/calcite-components-react@3.3.0-next.47) (2025-07-24)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.3.0-next.46](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.3.0-next.45...@esri/calcite-components-react@3.3.0-next.46) (2025-07-24)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.3.0-next.45](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.3.0-next.44...@esri/calcite-components-react@3.3.0-next.45) (2025-07-23)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.3.0-next.44](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.3.0-next.43...@esri/calcite-components-react@3.3.0-next.44) (2025-07-23)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
 ## [3.3.0-next.43](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.3.0-next.42...@esri/calcite-components-react@3.3.0-next.43) (2025-07-22)
 
 **Note:** Version bump only for package @esri/calcite-components-react